restore eager load logic

+++ dephell/commands/autocomplete.py
@@ -1,6 +1,6 @@
 # built-in
 from argparse import ArgumentParser
-from pathlib import Path
+from pathlib import Path, PurePath
 from platform import platform
 import os
 
@@ -53,23 +53,34 @@ class AutocompleteCommand(BaseCommand):
         # Install completions to the correct location for modern bash-completion.
         # This will be sourced on-demand by bash-completion as soon as dephell is
         # completed for the first time.
-        if 'BASH_COMPLETION_USER_DIR' in os.environ:
-            bashcomp_user_dir = Path(os.environ.get('BASH_COMPLETION_USER_DIR'))
-        else:
-            bashcomp_user_dir = Path(os.getenv('XDG_DATA_HOME', '~/.local/share')).expanduser() / 'bash-completion'
-        path = bashcomp_user_dir / 'completions' / 'dephell'
+        # https://github.com/dephell/dephell/pull/132
+        lazy_paths = (
+            Path(os.getenv('BASH_COMPLETION_USER_DIR', '')) / 'completions',
+            Path(os.getenv('XDG_DATA_HOME', '')) / 'bash-completion' / 'completions',
+            Path.home() / '.local' / 'share' / 'bash-completion' / 'completions',
+        )
 
-        path.parent.mkdir(parents=True, exist_ok=True)
-        path.write_text(script)
+        for path in lazy_paths:
+            if path.exists():
+                (path / 'dephell').write_text(script)
+                return
 
-        # We cannot reliably assume bash-completion 2.x exists on macOS,
-        # so inject it into the user's bashrc
-        if platform().lower() == 'darwin':
+        # https://github.com/dephell/dephell/pull/62
+        eager_paths = (
+            Path.home() / '.local' / 'etc' / 'bash-completion.d',
+            # ref. https://itnext.io/programmable-completion-for-bash-on-macos-f81a0103080b
+            Path('/') / 'usr' / 'local' / 'etc' / 'bash-completion.d',
+        )
+        for path in eager_paths:
+            if not path.exists():
+                continue
+            path = path / 'dephell'
+            path.write_text(script)
             for rc_name in ('.bashrc', '.profile', '.bash_profile'):
                 rc_path = Path.home() / rc_name
                 if not rc_path.exists():
                     continue
-                if 'completions/dephell' not in rc_path.read_text():
+                if str(path) not in rc_path.read_text():
                     with rc_path.open('a') as stream:
                         stream.write('\n\nsource "{}"\n'.format(str(path)))
                 break
